implemented attribute name map

+++ tests/ASolrDataProviderTest.php
@@ -58,7 +58,9 @@ class ASolrDataProviderTest extends CTestCase {
 	public function tearDown() {
 		foreach($this->fixtureData() as $attributes) {
 			$doc = ASolrDocument::model()->findByAttributes($attributes);
-			$this->assertTrue($doc->delete());
+            if (is_object($doc)) {
+			    $this->assertTrue($doc->delete());
+            }
 		}
 		$this->getConnection()->commit();
 	}
